Add files via upload
This is a project that I have signed NDA for. I do not have legal rights to disclose much information but here is the website for the company which is building our product. This is a cost-effective lice treatment device with a 95% success rate, cutting treatment time to 20 minutes. Focused on user safety and comfort by optimizing thermal design and control. What I offer to companies is a proven product development and CAD skills with a strong user-centered and safety-first approach—ready to help companies deliver innovative and reliable solutions.